Conflicts:
	modules/weblounge-contentrepository/src/main/java/ch/entwine/weblounge/contentrepository/impl/AbstractWritableContentRepository.java

+++ modules/weblounge-contentrepository/src/main/java/ch/entwine/weblounge/contentrepository/impl/PreviewGeneratorWorker.java
@@ -53,7 +53,7 @@ import java.util.List;
 class PreviewGeneratorWorker implements Runnable {
 
   /** The content repository */
-  private AbstractWritableContentRepository contentRepository = null;
+  private AbstractContentRepository contentRepository = null;
 
   /** The resource to render */
   private Resource<?> resource = null;
@@ -83,7 +83,7 @@ class PreviewGeneratorWorker implements Runnable {
    * @param styles
    *          the image styles
    */
-  public PreviewGeneratorWorker(AbstractWritableContentRepository repository,
+  public PreviewGeneratorWorker(AbstractContentRepository repository,
       Resource<?> resource, Environment environment, List<Language> languages,
       List<ImageStyle> styles, String format) {
     if (languages == null || languages.size() == 0)
@@ -158,6 +158,8 @@ class PreviewGeneratorWorker implements Runnable {
           return;
         }
 
+        long lastModified = ResourceUtils.getModificationDate(resource, l).getTime();
+
         // Create the remaining styles
         for (ImageStyle style : styles) {
 
@@ -168,10 +170,16 @@ class PreviewGeneratorWorker implements Runnable {
           FileInputStream fis = null;
           FileOutputStream fos = null;
           try {
-            fis = new FileInputStream(originalPreview);
             File scaledFile = ImageStyleUtils.createScaledFile(resource, l, style);
-            fos = new FileOutputStream(scaledFile);
-            imagePreviewGenerator.createPreview(originalPreview, environment, l, style, resourceType, fis, fos);
+
+            // Create the file if it doesn't exist or if it is out dated. Note
+            // that the last modified date of a file has a precision of seconds
+            if (!scaledFile.isFile() || FileUtils.isFileOlder(scaledFile, new Date(lastModified))) {
+              fis = new FileInputStream(originalPreview);
+              fos = new FileOutputStream(scaledFile);
+              imagePreviewGenerator.createPreview(originalPreview, environment, l, style, resourceType, fis, fos);
+            }
+
           } catch (Throwable t) {
             AbstractWritableContentRepository.logger.error("Error scaling {}: {}", originalPreview, t.getMessage());
             continue;
